+++ tests/__init__.py
@@ -1,1 +1,2 @@
 from .test_autoscheduler import TestAutoScheduler
+from .test_divideresults import TestDivideResults